+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The *alited* provides the rich and comprehensive set of features, some are uniqu
 # LAST CHANGES
 
 
-Version `v1.8.6 (17 Jul'24)`
+Version `v1.8.6 (20 Jul'24)`
 
   - BUGFIX: Ctrl+N, Alt+H => bad text index ".0"
   - BUGFIX: e_menu's bug injected in v1.8.5
@@ -21,7 +21,7 @@ Version `v1.8.6 (17 Jul'24)`
   - CHANGE: tip of status bar's right info
   - CHANGE: Tools/Run...: appearance
   - CHANGE: docs
-  - CHANGE: packages: e_menu/apave 4.4.7
+  - CHANGE: packages: e_menu/apave 4.4.8
 
 
 Version `v1.8.5 (3 Jul'24)`